Convert OvmsCommandMap key from std::string to const char*
This change reduces the size of the OvmsCommand::Execute() stack frame
from 96 to 32, the minimum stack size.

In a test where at startup the initial SIMCOM code is doing
CheckPowerOff and a startup script is starting wifi client and telnet,
and then a script is run to display the m.freeram metric, the free
space is increased from 9156 to 26016.