Check for empty LogBuffers before doing output
1. In Output(), if there has been nothing output to the LogBuffers
that stores output of the BufferedShell, don't bother sending it to
the OvmsWriter.

2. If a nested script outputs to BufferedShell::Log() but the
LogBuffers in the BufferedShell is empty, we would crash while trying
to split it in.  Now just take the LogBuffers supplied as an argument
to Log() and put it in place of the empty one that BufferedShell had.