Revise memory diagnostics to use task handle and full task name
NOTE!  This code change depends upon changes in the memory debugging
code in esp-idf to track ownership of malloc blocks using a task
handle rather than three characters of the task name.

This change allows referencing tasks by their full name so it is not
necessary to make the names unique in three characters.  It also
allows referencing specific instances of a task in cases where there
may be more than one.

In order to translate task names to task handles and vice versa, this
change implements a bidirectional TaskMap (not using std::map).  To
populate that map, an AddTaskToMap() function was defined and is
called in a few places where tasks are created.  The map is also
populated with the current list of tasks whenever one of the memory
diagnostic commands is issued.  As a consequence, task diagnostics and
memory diagnostics are tied together.

One penalty of this change is that the "module abort" command can't be
set up to trigger on allocations of a task before that task is
created.  A workaround is to call mem_malloc_set_abort() directly
after the task creation.  Another is to set the trigger to not be
specific to a task, just on allocation size and number of allocations.

Storage of arrays for the memory diagnostics is no longer allocated
statically nor on the stack, it is all allocated dynamically from IRAM
that only allows 32-bit accesses.  That leaves all the 8-bit
accessible heap available for production code and reduces the required
allocation of stack space.

Some specific details:

- The "module tasks" command now displays the total 8-bit-accessible
  heap allocation for each task as another column.

- Aggregation of the sizes of blocks allocated by each task to get a
  total is now done in the OS so it is not necessary to dump info for
  each individual heap block into an array for the app code.  Only the
  blocks belonging to specific tasks will be dumped when those tasks
  are selected by the user for a detailed list of blocks.

- Removed extra prefix letters in telnet tasks that had been added to
  make them unique in three characters.

- The conditionals to check for the necessary esp-idf configuration
  have all been gathered to the top of the ovms_module.cpp file
  rather than being spread through the commands.