Fix two bugs in WolfSSH discovered in initial testing
The first bug is only exposed if there's not enough free heap to
allocate the WOLFSSH structure.

The second bug was a copy/paste error that is only exposed if a user
supplied recv function is configured while retaining the built-in send
function.

At this point neither of these conditions applies, so these bugs would
not be hit anyway.

These bugs have been reported to Wolf.