Restore the standard high-water mark of 1GB
The SSH high-water mark is the number of bytes to transmit after which
rekeying is required.  That value is limited by the protocol to no
more than 1GB.  We set it to the default WolfSSH value of 1GB - 32KB.
In practice this probably means we never hit it, which is good because
ConsoleSSH does not implement rekeying yet.