Modify mg_send() to actually transmit if the socket is ready
Use a no-wait select() call to determine if the socket is ready to
transmit, and if so, transmit the data that has been buffered.  This
will usually not block, but can block if the amount of data buffered
before the socket became ready is large.

This change is needed for the Telnet and SSH consoles because
otherwise with a large output mg_send() could consume all available
dynamic memory and therefore cause various failures such as the SSH
connection to hang.