Implement feature for commands to take input from console
A command that wants to take input from the console should do the
following:

1. Allocate memory to hold state information it needs to maintain
during its execution.

2. Call writer->RegisterInsertCallback(InsertCallback, void*) where
InsertCallback is:

bool (*InsertCallback)(OvmsWriter* writer, void* userData, char);

3. Return.

4. In its callback function, it should process the provided input
character and return true if more input is desired or false if the
command is finished.  The processing must include echoing the
character if appropriate.

See the example command "echo" in ovms_command.cpp.