Change policy to always discard log messages if queue full
Since delaying tasks may cause problems worse than not being able to
see a log message from the task, we will just discard a log message if
the queue is full when the message is issued.

Lost messages will be counted and when the queue eventually empties a
message will be printed telling the number of log messages lost.

Note: We hope it is unlikely that the total number of messages lost
between reboots never reaches 2^32, but it does not matter if the
counter of lost messages wraps around because we use modulo 2^32
arithmentic to calculate the number of messages lost but not yet
acknowledged as lost.  We assume that difference will never approach
2^32.  Using two separate counters this way avoids race conditions.