Revisions for compatibility with esp-idf release/v3.0
Mostly of this change is in revisions to ovms_module.cpp for
compatibility with a new implementation of the OS code supporting the
"module memory" and "module task" commands.  This includes a change to
a menuconfig option in the OS rather than in OVMS, so the previous
options are removed from OVMS.  Also the "module abort" command is
removed because the necessary OS support is no longer available and
the new heap tracing functions are a better replacement.

In addition there are various other changes required to compile with
release/v3.0 of esp-idf:

- In esp32adc.cpp, adc1_get_voltage() -> adc1_get_raw().

- In simcom.cpp and console_async.cpp, the UART config struct gets a
  new member use_ref_tick.

- Switch spi_master_nodma.c and ovms_housekeeping.cpp to permanently
  use the new names for the heap caps header and API (previously there
  commented out).