ESP IDF 3.x changed the way file descriptors are implemented.
The IDF 3.x now uses an offset to separate ranges of file descriptors, so the
responsible subsystem can be identified (LWIP, VFS files, etc). Kind of a kludgy
solution, and ends up with sockets being numbered 8192 onwards. Then, they have
code to specially handle FD_SET and remove the offset. Horrible solution.

Using sys/select.h is a big NO NO with this, as we will use the wrong FS_SET kludge,
so Espressif use the ESP_PLATFORM to disable this in newlib/include/sys/types.h.

In our case, and the problem here, we were using:
  int result = select(FD_SETSIZE, &fds, 0, 0, &timeout);
Here, FD_SETSIZE is the FDs without offsets, but what we have to pass is
the biggest FD with offsets (first parameter). Workaround approach is:
  int result = select((int) sock + 1, &fds, 0, 0, &timeout);

Horrible. But it works.