Additions to the usage text for 'module memory' command
Argument '*' gathers individual block records for all tasks up to the
limit of the array we provide to the OS and then prints in the second
section of the output all the changed blocks belonging to tasks that
have changed totals in the first section of the output.  Since there
may be too many blocks to capture them all, individual task names or
addresses can be listed as arguments instead.

Argument '=' causes the first section of the output to list all tasks
including those whose total allocations have not changed from the
previous execution of the command.  Individual block records are not
gathered (unless the source code is changed to populate ctasks[]).