Override wolfssh and wolfssl malloc functions to log failures
Use built-in mechanisms in wolfssh and wolfssh (wolfcrypt) to override
the malloc, free and realloc functions with our own functions that log
an error message if the allocation fails.  This is to avoid confusion
when wolfssh reports a higher-level error, such as WS_CRYPTO_FAILED,
when the problem is really caused by memory allocation failure.