Stop the chargen output loop if write() returns an error
The "test chargen" command outputs 1000 lines by default.  For an ssh
server running on OVMS v3.0 hardware, this will exceed the buffer
capacity in mongoose and the socket layer due to limited RAM.  When
the limit is reached, write() will return an error that we can detect
and stop the loop.