Framework: provide InternalRamAllocated class & methods    (counterpart of ExternalRamAllocated)
This is meant to ensure heap placement of objects that need to be
in internal RAM, like IRAM interrupt handler data or generally RAM
for functions that must not be blocked by an SPIRAM cache operation.

To apply, add InternalRamAllocated as a super class.

With CONFIG_OVMS_HW_SPIMEM_AGGRESSIVE enabled (default), this forces
creation of class instances in internal RAM.
With CONFIG_OVMS_HW_SPIMEM_AGGRESSIVE disabled, this changes nothing.