#162 Fix for 0% SOH on 30KWh models with Nissan BMS update
Tuns out the PollReply for the battery info is 41 bytes long on these models instead of the 39 bytes we have previously seen on the 24KWh models.
The contents still seem to have the same structure as before, it is just that 2 unknown bytes are added to the end of the reply.