Notifications via WebSocket channel & Web UI dialog widgets
https://github.com/openvehicles/Open-Vehicle-Monitoring-System-3/issues/83

Includes addition of a fifth notification channel "stream" for subtype/app
specific data. Notification subtypes are used as MQTT style topics:

  "notify/<type>/<subtype>"

…with subtype dots '.' replaced by '/'.

The web client subscribes to topics on a receiver object. Types "info",
"error" and "alert" are always forwarded to the web client, "data" and
"stream" only as subscribed. MQTT topic patterns are supported (as far as
mongoose implements them, which is currently only "#").

Test & documentation:
- components/ovms_webserver/dev/notifications.htm
- components/ovms_webserver/dev/dialogtest.htm