Consolidate print operations
This is a follow-on to commit 74169ce67750332b69d2b577ca3040dcbded7f07.

With SSH, each print() call in microrl was transmitted as a separate
packet, so the many small print operations made output slow.  This
commit consolidates the printing of the token list for tab completion
into a buffer rather than printing each token separately.  This was
particularly slow for the "power" command.