Add "module leaks" subcommand
New subcommand "leaks" is a variant of "memory" for use when
specifying one or more task names to display the blocks allocated to
those tasks.  The "leaks" variant does not display the possibly many
blocks that were retained from the previous instance of the command,
only those that have been freed or newly allocated; a count of those
blocks is shown instead.  This makes a more compact display that is
expected to be useful when investigating memory leaks.

In print_blocks(), variable "name" is now set once at the top of the
function rather than being set unnecessarily inside loops (since the
code has evolved).