Remove unsafe OvmsCommand::m_usage, GetUsage -> PutUsage
Since commit 00451051496887dfe03d028ce0d64111cc76cc76 the usage string
for a given OvmsCommand object is no longer constant so it can't be
cached.  But we were still using the member m_usage, which used to
hold the cached value, as the string in which the usage message was
constructed.  This is not safe if multiple command executions might be
running at the same time.  Also, there was a bug that the m_usage in a
child command would not be cleared before being constructed, resulting
in replications.

To get rid of m_usage and construct the string in a local variable
instead required changing from GetUsage returning m_usage.c_str()
(can't return c_str() on a local variable) to PutUsage that outputs to
the writer before returning.

This change also reduced the stack depth by 16 bytes.