Move GetCompletion() from OvmsWriter into OvmsCommandMap
After tab completion of command words was implemented in commit
7c9689aae5446ace9721154a3e48425b30c80989, Mark asked why
GetCompletion() was implemented ConsoleAsync (under OvmsWriter) rather
than in OvmsCommandApp since the latter has the necessary information.
The reason is that storage for the array of completion tokens needed
to be in the ConsoleAsync object because that is where the state
associated with that command parser session is located and
OvmsCommandApp is shared among all command parsers.

Now I realize that it's more appropriate to move the loop searching
the command words into a new GetCompletion() in OvmsCommandMap where
those words live, and have it reach back to the writer using a reduced
method SetCompletion() just to store the results.  Also implement a
NOP method in OvmsWriter itself, rather than a pure virtual method, so
other derived classes don't have to implement the NOP method.

This is in preparation for adding tab completion for command
parameters, which Mark was also requesting at that time.