Allow intermediate params in commands and param tab completion
This commit adds yet another parameter to the OvmsCommand constructor
with corresponding member: a function to validate the parameters
required for that OvmsCommand.  This is to allow parameters to be
collected in the middle of a command syntax with command tokens after
the parameters.  That OvmsCommand would typically not have an execute
function; instead OvmsCommand::Execute() will call the validate
function to check the validity of the intermediate parameters in the
command sequence and if valid then proceed to the next child
OvmsCommand selected by the token following to intermediate
parameters.

If the validate function is supplied for an OvmsCommand object it will
also called when a TAB character is entered within the input of the
parameters for that object.  If appropriate, the validate function can
provide the set of matching values so that microrl can display them on
the console.