Allow access to intermediate parameters for execution
Save the argv pointer from the top-level command into the OvmsShell
object to allow the execution function of the lowest-level subcommand
to access the parameter values of intermediate subcommands.  The argv
would indexed by counting down the number of subcommand levels from
the top.

To allow for cases where the intermediate parameters may sit at a
variable number of subcommand levels from the top-level command but a
fixed number of subcommand levels up from the lowest, the argv array
can be indexed negatively from the argv parameter supplied to the
lowest subcommand.  This required not setting that argv parameter to
NULL when the accomanying argc is zero even though argv would then be
pointing past the end of the array.  That's not really a risk because
argc==0 means that argv[0] should not be accessed.