Revise RegisterCommand arg defaults, update all calls
Very few instances of RegisterCommand() took advantage of the argument
defaults because the last one, 'secure', defaulted to false when most
instances needed true.  So this commit flips that default.

In addition, the default for the 'max' arg was INT_MAX whereas most
instances should have max = 0, so that default is also changed.  And
lastly, a new default value NULL was added for the 'execute' function
pointer.

All calls of RegisterCommand were changed to make maximal use of the
arg defaults so that the existing code can serve as a useful pattern
for the addition of new commands.  In particular, for secure,
non-terminal commands, such as the top-level "framework" commands, the
model should simply be:

    RegisterCommand("name", "Title");

For secure, terminal (sub)commands that don't require any additional
parameters, the model should be:

    RegisterCommand("name", "Title", execute);

This model also applies if the command has children but the command
itself wants to execute a default operation if no subcommand is
specified.  It is incorrect to specify min = 0, max = 1 to indicate an
optional subcommand; that is indicated by the presence of the execute
function at the same time as a non-empty children array.

Any command with required or optional parameters should provide a
usage string hinting about the parameters in addition to specifying
the min and max number of parameters:

    RegisterCommand("name", "Title", execute, "usage", min, max);

The iMiEV and Kia Soul command hierarchies marked the top-level
command as secure but all the subcommands as not secure.  This does
not make sense as the subcommands could not be reached if not
enabled.  Furthermore, a subcommand to unlock the doors should be
secure indeed.  This commit changes all the subcommands to be secure.

Similarly, the 'status' subcommands for 'server v2' and 'server v3'
were marked not secure, apparently with the intention of letting the
status be examined without enabling, but the top-level commands were
marked secure so the status command could not be reached.  The status
subcommands are now marked secure.