CAN LOGGING framework rework
A major set of extensions (and basic re-write) of the CAN logging framework.

The system is now split into three:
  1] The loggging framework itself
  2] The formatters (to convert to/from formats like pcap, crtd, etc)
  3] The loggers (to read/write to different locations such as monitor, vfs, tcp, etc)

Also standardised are the formatters (now shared between can logging and retools),
and the filters (so software filters can be defined and handled in a standard way).

The tcpclient and tcpserver loggers have not yet been implemented. They are there
purely stubs at the moment.

This is a very large change, and has not had extensive testing yet. There will
likely be a number of changes in the coming few days / weeks.