Network: workaround against data loss on poor wifi connection
The wifi blob does a bad job at detecting when to leave a network,
leading to message losses when the signal quality is on the edge.
To circumvent this, this change delays using the wifi net until
the signal is good, and disconnects as soon as the signal gets bad.

New configs:
  [network] wifi.sq.good  [dBm], default -87
  [network] wifi.sq.bad   [dBm], default -89

New events:
  network.wifi.sta.good
  network.wifi.sta.bad