Metric ms_v_bat_soh uses "stale-threshold" SM_STALE_HIGH. ms_v_bat_voltage is a closely related metric, but was SM_STALE_MID.
This means that when your car is asleep the voltage metric expires
but the soh does not.

This impacts the dashboard.

For me it makes since that of the SOH is SM_STALE_HIGH then the
voltage should be the same since the SOH fundamental is derived from
the pack voltage.