Avoid defining XREALLOC for FREERTOS if XMALLOC_USER is defined
This appears to be a bug in the maze of configuration settings for
WolfSSL.  I have submitted a support request to WolfSSL about it.

We need to define XMALLOC_USER along with the desired implementations
of XMALLOC, XFREE and XREALLOC so that we can install our own
functions as callbacks for the three operations.  Without this change
there is a compilation conflict for XREALLOC.