Cellular: try to avoid duplicate system.modem.down events raised
It seems we have a race condition in ovms_netmanager PrioritiseAndIndicate when it accesses the netif linked
list without locking, if interfaces are added/removed during traversal of the list. We should really lock
this list, but it is not easy, as part of LWIP code and there doesnt seem to be any external API to that.