TR: enable manually starting charge during scheduled wait
If a Tesla Roadster is asleep while waiting for a scheduled charge
time, attempting to manually start a charge from the app would fail.
Before sending the CAN message to start the charge it is necessary to
send a "wakeup" CAN message.  As the car wakes up it temporarily
indicates that there is no pilot signal even though the cable is
connected.  This causes the app to hide the charge connector icon.
To fix this we add a new state variable to track the steps the car
goes though as it wakes up, and we fake the status reported to the app
to indicate that the car is in "prepare" state instead.