TR 2.x: enable manually starting charge during scheduled wait
Followup to the previous commit to handle differences in the charging
state as reported by 2.x cars when attempting to manually start a
charge from the app while the car is asleep and waiting for a
scheduled charge time.

Only reset the m_starting_charge state back to INACTIVE when charging
starts or is interrupted before starting.  The state was being reset
when it changed to "preparing".