TR: improve robustness for manually starting charge from app
Don't change the m_starting_charge state when a CAN status message
indicates charging has been stopped by request because we may see this
message when we are trying to restart charging from the app when the
car was sleeping.  The state will already have been set back to
INACTIVE when charging started.

Rather than using the charge substate to determine conditions in which
we may need to issue a repeated start command after the car wakes up,
do this whenever we are starting the charge while the car is asleep.

More generally, avoid undesired changes to the m_starting_charge state
because a status message is received and interrupts the execution of
CommandStartCharge() between the wakeup command and the start command.
We choose the desired next state at the top of the function but set it
just before returning.