TR: Remove workaround sending "stopped" for "timerwait" to app
Because the iOS app would not send a start-charging command when
requested with the charge connector slider if the Tesla Roadster was
in "timerwait" state, a workaround was put in OVMS firmware to
indicate the "stopped" state instead.  But that caused the app to
generate false "charging stopped" notifications as the charge cable
was plugged in.

Now the iOS app has been fixed, so this workaround is removed.