fix negative integer DBC numbers having wrong values
When non-32-bits negative numbers were processed by the DBC processor, there
was an issue where the resulting metric values were incorrect.

As an example:

A 16-bit signed value coming from the can bus as 0xFFF7(-9) is being multiplied
by a factor of -1.
When handled by the DBC processor, the incoming number is stored as 0x0000FFF7 ; which is
not -9, but 65527.
After multiplication, it comes as 0xFFFF0009(-65527).

(While those should be: 0xFFFFFFF7(-9) and 0x00000009(9))

The root cause is that the sign of the non-32-bit number is not extended to the left,
thus making it a different number.

The fix is to sign-extend the number to the left (it should work from 2-bit numbers
to 32-bit numbers) when we're dealing with signed numbers.

Co-Authored-By: Michael Geddes <frog@bunyip.wheelycreek.net>
Signed-off-by: Ludovic LANGE <llange@users.noreply.github.com>