small change to the way the metrics listener are registered (use std::string instead of const char * for first parameter)
Previously the metrics listeners were registered by storing the `caller` parameter
as a `const char *`.
This made mistakes possible when we would register and deregister with the same string
but not the same memory address.

This patch changes the `caller` parameter type to `std::string` which will prevent this kind
of mistakes.

Signed-off-by: Ludovic LANGE <llange@users.noreply.github.com>