enables configuration of events that will be logged to can logs
A new configuration item `can log.events_filters` allows to choose
which events will be sent to the can logs.

The format is a comma-separated list of events filters, possibly
containing wildcards (`*`).

Any filter starting with `*` is an `endsWith()` comparison.
A filter ending with `*` is a `startsWith()` comparison.
A filter without any wildcard is a `==` comparison.
(A wildcard in any other position in the filter is invalid)

To maintain compatibility, the default value is `x*,vehicle*`, which
is going to allow only the events starting with `x` or with `vehicle`.

The default value is equivalent to entering the following command:
```
config set can log.events_filters "x*,vehicle*"
```

Signed-off-by: Ludovic LANGE <llange@users.noreply.github.com>