CMake : workaround for ovms_webserver defines
This change is there to have consistency between the ESP-IDF v3 build system, and the ESP-IDF v5 build system.

When porting `ovms_webserver` component makefile to CMake we encountered an issue ; we
were not able to reproduce this set of defines based on generated files.

The workaround is to generate a header file (instead of the defines), and to
have an helper shell do the generation.

We may revert when we have found a way to make it work with CMake.

Signed-off-by: Ludovic LANGE <llange@users.noreply.github.com>