ESP-IDFv4+ / CMake : convert wolfssh to a git submodule
While not absolutely necessary, it seemed easier to manage the compilation recipes (mainly `component.mk`, `CMakeLists.txt`)
by moving the `wolfssh` component one directory level below where it sits at the moment.
This way, we can have our own files decoupled from the whole external project sources.

We could have stopped here, but also converted the external project sources into a git submodule.
It should help us by easing the upgrades to a new upstream release (just by changing the submodule
pointer) - however as a downside it's more complex to add ad-hoc patches to the repository.
(We will have to convince upstream that our patch is interesting, which could slow the process a bit, and in the meantime
we can use a fork - like we do for `mongoose`)

By luck `wolfssh` had not (yet) been patched by us, so we're not loosing anything.

As submodules are already used in the project, it made sense to use a similar mechanism to track
external projects.
(Other options could be to: not change anything ; use git subtrees ; or event Google's `repo` tool.)

Signed-off-by: Ludovic LANGE <llange@users.noreply.github.com>