ESP-IDF v5+: update WatchDogTimer initialization
We introduced the dependency on the (official) configuration item `CONFIG_ESP_TASK_WDT`
to decide whether (or not) we want to initialize the WDT.
In ESP-IDF v5+, it's not possible any more to (re)initialize and already initialized timer,
and if `CONFIG_ESP_TASK_WDT` the WDT would have been initialized already during boot.

Note: For some reason, in ESP-IDF v5+, we need to define `CONFIG_ESP_TASK_WDT=n` otherwise
it won't work as expected.

Signed-off-by: Ludovic LANGE <llange@users.noreply.github.com>