ESP-IDF v4+: we do not have uxMutexesHeld any more
In ESP-IDF 3.3.x builds of OVMS, we're using a fork of ESP-IDF. In this fork,
FreeRTOS has been patched with openvehicles/esp-idf@95e43fc2c4360f8126a0985bf037a293bebe3767 to
add a `uxMutexesHeld` field in `TaskStatus_t`.

However, in the ESP-IDF >= 4 build we're using mainstream ESP-IDF (for the moment),
which does not include this field.

So we adapt the display to handle both cases.

Signed-off-by: Ludovic LANGE <llange@users.noreply.github.com>