ESP-IDF4+ : `esp_task_wdt_get_trigger_tasknames()` not available
`esp_task_wdt_get_trigger_tasknames()` was a custom patch of our ESP-IDF 3.3.4 fork.
It is used to list the names of the task(s) responsible for triggering the WDT,
and this information is saved in the (persistent) boot_data structure for post-crash
analysis.

Unfortunately it's not available on upstream ESP-IDF 4+.
For the moment we remove the feature, until we find a way to implement it without
patching ESP-IDF 4+.

Signed-off-by: Ludovic LANGE <llange@users.noreply.github.com>