identify dependencies to mongoose, and generate a mongoose version header
In preparation of a future mongoose upgrade, we need to detect current
mongoose version.
This is done by generating a private header `include/mg_version.h` which is
generated at compile time (both by `component.mk` and `CMakeLists.txt`, from
a template file `mg_version.h.in` and the version in `mongoose.h`)

To ensure proper compilation order (the generated file needs to be created before
being included), we need to identify components that depend on the build of mongoose.
This is recorded both in `component.mk` and `CMakeLists.txt`).

Signed-off-by: Ludovic LANGE <llange@users.noreply.github.com>