Update README.md - update description to represent hardware version 3.3
The current hardware version 3.3 is the one with the highest cost. Many people report on the internet that they do not get the OVMS because of the high cost. Its not that affordable for many people so just remove the comment about the price because it does not add any benefits. The people see the price at the distributor and decide on their own if they can afford it. Hardware version 3.3 have dual certification and the modem can be used worldwide. Also it supports 2g, 3g and 4g networking.