Sync compression level for gz web assets and remove from repo
The CMake build and the legacy make build do not generate the same gzip
files, due to different compression settings, so the worktree is getting
dirty as soon as it's configured.

While fixing this by using the higher compression also for make, take
the opportunity to remove the compressed files from the repo. The
intermediate concatenated files are kept to support the web UI dev env.