renaulttwizy: Fix deprecated operations on volatile
Extract the duplicated code to a helper method.

Do the flag setting via a local variable to make it explicit when the
volatile variable is accessed.

The sensors state really should use atomics if it may be accessed from
different contexts (threads/interrupts), or otherwise just remove
volatile since it's not doing anything useful.