ESP32CAN, MCP2515: add hardware acceptance filter APIs & shell commands
Note: the shell commands are meant as a development aid, thus intentionally
not added to the "changes.txt". Once determined, the proper filter
configuration should be done by the application/vehicle.

The "setaccfilter" command breaks the generalized CAN command syntax, as
the filters work fundamentally different on the ESP32CAN and the MCP2515.
See command help on can1 & can2 for details.