cellular -> NetLoss loop fix
When the modem detects a NetLoss, it usually enters a continuous loop of NetLoss, NetWait and NetStart when a Wi-Fi connection is used. The problem can be simulated with ‘cellular setstate NetLoss’. The fix provides that when a NetLoss loop is detected, the modem then performs a PowerOffON, which reliably solves the problem.