Update vehicle_thinkcity.cpp
TPMS Status (0x345) Case added.  This addition uses Standard Metrics for TPMS Alerts and Warnings to allow TPMS Sensor Status data to be reflected in the TPMS GUI of the vehicle App as well as in the notifications.  These changes were included in a local build, tested on 2 THINK City vehicles and functionality verified.  
New Behavior:  On starting to drive, each tire in the TPMS GUI now shows a check mark if the tire sensor is confirmed working and tire pressure not very low.  Otherwise, that tire label comes up RED and with an Alert triangle.  After about a minute when first pressures come in, the triangles and check marks go away and are replaced by pressures.  The color assignments remain as long as alert conditions persist.  Testing confirmed the RED alert goes away a few minutes after correcting a low pressure condition.
Remaining issue: Left vs Right tire data always or often appears switched, however this is likely due to the Automatic TPMS Learning incorporated with THINK City cars.  Direct TPMS CAN data collection has shown that either the L/R tire order in the data is flipped or that it could go either way when the TPMS sensors are automatically "learned".  Until we know for sure which it is, no sense in changing anything.